feat: Add project management components including ProjectHeader, ProjectTabs, ProjectUsers, KanbanBoard, TaskCard, TaskColumn, and ProjectStats
- Implemented ProjectHeader for displaying project details and action buttons.
- Created ProjectTabs for navigation between Kanban, chat, and statistics views.
- Developed ProjectUsers for managing project user assignments with modal support.
- Introduced KanbanBoard for task management with drag-and-drop functionality.
- Added TaskCard for individual task representation within Kanban columns.
- Implemented TaskColumn to group tasks by status in the KanbanBoard.
- Created ProjectStats to visualize project task statistics and timelines using charts.
